requirements.txt
# ============================================
# GSLTS Trading System - Python Dependencies
# ============================================

# Core
python-dotenv==1.0.0

# HTTP & API
requests==2.31.0

# WebSocket
websocket-client==1.6.4

# Database
pymysql==1.1.0
sqlalchemy==2.0.23

# Cache
redis==5.0.1

# Backend Framework (Phase 2)
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Data Processing
pandas==2.1.3
numpy==1.26.2

# Scheduling
apscheduler==3.10.4

# System Monitoring
psutil==5.9.6

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
